==> FD-Analyses/Codes/7.-FD-Analyses.Rmd <==
---
title: "SHARKS FD"
author: "Catalina Pimiento"
date: "27/06/2022"
output:
  html_document:
    highlight: zenburn
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
---

# 1. Data treatment: select common species (traits, grids, phylo)

```{r chunk1, message=FALSE, warning=FALSE}
library(readxl)
library(stringr)
library(tidyverse)
library(dplyr)
library(ade4)
library(beepr)
library(raster)
library(rgdal)
library(readxl)
library(janitor)
library(tidyselect)

# SELECT SPECIES TO WORK WITH: w/ trait, plylo & spatial data and fully marine

# read trait data
  # traits.curated<-read_xlsx("~/Dropbox (Smithsonian)/SHARKS/Traits/shark_traits.xlsx") # original dataset
  # traits.new<-read_xlsx("~/Dropbox (Smithsonian)/SHARKS/Traits/shark_traits_corrected.xlsx") # curated dataset
    traits.curated<-read_xlsx("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_inputs/shark_traits_curated.xlsx")
  
# merge datasets  
  # traits.curated <-  traits.curated %>%
  #   dplyr::select(-habitat, -terrestriality, -vertical, -diet, -`max length (cm)`)%>%
  #   left_join(traits.new %>%
  #               dplyr::select(Species, habitat, terrestriality, vertical, diet, max.length),
  #             by="Species")

# species from which we don't have phylogentic data from
  no_phylo <- c("Aetobatus ocellatus", "Aetomylaeus milvus","Apristurus breviventralis" ,
                "Apristurus garricki" ,"Miroscyllium sheikoi", 
                
                "Apristurus nakayai", "Breviraja mouldi","Carcharhinus humani", 
                "Centrophorus uyato", "Hemiscyllium halmahera", "Himantura bleekeri", 
                "Pristiophorus lanae", "Rajella paucispinosa", "Tetronarce cowleyi")

# species strictly freshwater species    
  freshwater <- c("Potamotrygon falkneri","Potamotrygon schroederi", "Potamotrygon tigrina")
    
# species from which we have spatial data from:
  # load spatial data
    load("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/grids_biomes_no.syn.RData")
 
  # check differences and commonalities
   #  are there differences between trait species and grid species?
      length(setdiff(traits.curated$Species, colnames(grids[,4:dim(grids)[2]])))
      length(setdiff(colnames(grids[,4:dim(grids)[2]]),traits.curated$Species))
      
# get common species
  traits.curated<- traits.curated %>%
      dplyr::filter(!Species %in% no_phylo)%>%
      dplyr::filter(!Species %in% freshwater)
  # how many species do we have with both traits and spatial info?
    names(grids[1:5])
    length(intersect(colnames(grids[,4:dim(grids)[2]]),traits.curated$Species))
# find common species between traits and spatial data
  common.species <- intersect(colnames(grids[,4:dim(grids)[2]]),traits.curated$Species)
  length(common.species)
  
# filter grid dataset so that they only include common species
  grids.fd <- grids %>%
    unite("grid", lon:lat, remove = FALSE) %>%
    dplyr::select(-lon, -lat) %>%
    column_to_rownames("grid") %>%
    dplyr::select(common.species)
  
# select**GRIDS** with more than 4 species (space is 3D) 
  RS<-apply(grids.fd, 1, sum)
  #which(RS==0)
  #sum(grids.fd[9711,]) # zero species 
  grids.fd<-grids.f[-which(RS<4),]
  which(apply(grids.fd, 1, sum)<4)

# select species that are no longer present in any cell after removing those with less than 4 sp
  SumRS <- apply(grids.fd, 2,sum)
  rare <- names(which(SumRS==0))

  grids.fd <- grids.fd %>%
    dplyr::select(-rare)
  
  save(grids.fd,file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/grids_commonspecies.RData')
 
# select trait species with only common species and those with enough grid data  
  traits  <- traits.curated %>%
    dplyr::filter(Species %in% common.species) %>%
     dplyr::filter(!Species %in% rare)
  
  write.csv(traits, file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/shark.traits.csv')
  
  setdiff(names(grids.fd), traits$Species); setdiff(traits$Species, names(grids.fd))
  dim(grids.fd); dim(traits)

 working.species <- traits %>%
       dplyr::select(Species)
 
 write.csv(working.species, file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/working.species.csv')

```

# 2. Trait analyses

```{r chunk2, message=FALSE, warning=FALSE}   
# get modal values from imputed data: joint diet and IUCN  
  load('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/imp.mode.RData')

# add traits that were not imputed and select species we will be working with
  shark.traits <- read_csv("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Imputations/Oct_2021/shark.traits.imputations_all2.csv") 
  
  traits_imp <- imp.mode %>%
    bind_cols(vertical.benthic=shark.traits$Vertical.benthic, vertical.pelagic=shark.traits$Vertical.pelagic,
              thermo=shark.traits$thermo, feeding=shark.traits$feeding) %>%
    filter(Species %in% traits$Species)
  
  dim(traits_imp)
  
  traits_imp <- as.data.frame(traits_imp)

# prep trait data
  
  habitat <- traits_imp %>%
    dplyr::select(Species, habitat.coastal, habitat.oceanic)
    rownames(habitat) <-  habitat$Species; habitat$Species <- NULL
    habitat <-prep.binary(habitat, col.blocks = 2, label = "habitat")

  vertical<- traits_imp %>%
    dplyr::select(Species, vertical.benthic, vertical.pelagic)
    rownames(vertical) <-  vertical$Species; vertical$Species <- NULL
    vertical <-prep.binary(vertical, col.blocks = 2, label = "vertical")
  
  terrestriality<- traits_imp %>%
    dplyr::select(Species, terrestriality)
    rownames(terrestriality) <-  terrestriality$Species; terrestriality$Species <- NULL
  
  thermo<- traits_imp %>%
    dplyr::select(Species, thermo)
    rownames(thermo) <- thermo$Species;thermo$Species <- NULL

  feeding<- traits_imp %>%
    dplyr::select(Species, feeding) 
    rownames(feeding) <- feeding$Species;feeding$Species <- NULL

# diet
 diet<- traits_imp %>%
    dplyr::select(Species, diet.plankton:diet.highverts)
    rownames(diet) <-  diet$Species; diet$Species <- NULL
    diet <-prep.binary(diet, col.blocks = 4, label = "diet")
      
 # TL 
   size<- traits_imp %>%
    dplyr::select(Species, size) %>%
    mutate(log.size = log10(as.numeric(size))) %>%
    dplyr::select(-size)
 	  size<-as.data.frame(size)
    rownames(size)<-size$Species; size$Species <- NULL
    
# prepare the data for the dist.ktabl function that generalize the Gower's distance
  ktab1 <- ktab.list.df(list(habitat, vertical, terrestriality, thermo, feeding, diet, size))
  
# create distance matrix
  distrait <- dist.ktab(ktab1, c("B", "B", "O", "O", "O", "B", "Q"), c("scaledBYrange"))
  distrait<-distrait^2
  save(distrait,file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/distrait.RData')
  #load('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/distrait.RData')
  length(attributes(distrait)$Labels)
```

# 3. Traits across clades

```{r chunk3, message=FALSE, warning=FALSE}
library(wesanderson)
library(cowplot)
library(grid)
library(visreg)
library(ggcorrplot)

# convert traits as factors
  traits_factors <- traits_imp %>%
                     mutate(habitat=1) %>%
                     mutate(habitat = replace(habitat, habitat.oceanic==1&habitat.coastal==1, 2))%>%
                     mutate(habitat = replace(habitat, habitat.oceanic==1&habitat.coastal==0, 3))%>%
                    
                     mutate(diet=1)%>%
                     mutate(diet=replace(diet, diet.plankton==1&diet.inverts==0&diet.fish==1&diet.highverts==0, 2)) %>%
                     mutate(diet=replace(diet, diet.plankton==0&diet.inverts==1&diet.fish==0&diet.highverts==0, 3)) %>%
                     mutate(diet=replace(diet, diet.plankton==0&diet.inverts==1&diet.fish==1&diet.highverts==0, 4)) %>%
                     mutate(diet=replace(diet, diet.plankton==0&diet.inverts==0&diet.fish==1&diet.highverts==0, 5)) %>%
                     mutate(diet=replace(diet, diet.plankton==0&diet.inverts==1&diet.fish==1&diet.highverts==1, 6)) %>%


                     mutate(vertical=1) %>%
                     mutate(vertical = replace(vertical, vertical.benthic==1&vertical.pelagic==1, 2))%>%
                     mutate(vertical = replace(vertical, vertical.pelagic==1&vertical.benthic==0, 3))%>%
    
                     mutate(iucn_cat = recode(iucn, "0"="LC", "1"="NT", "2"="VU", "3"="EN","4"="CR")) %>%
                     mutate(iucn_cat2 = recode(iucn, "0"="NoT", "1"="NoT", "2"="AT", "3"="AT","4"="AT")) %>%
                     dplyr::select(Species, habitat, terrestriality, diet,  vertical, thermo, feeding, size, iucn_cat, iucn_cat2) %>%
                     left_join((traits %>% dplyr::select(Species, Order, Superorder)),
                     by="Species") %>%
                     mutate(across(c(habitat:diet,vertical:feeding), factor))
                   
     traits_factors$size <- log(as.numeric(traits_factors$size))

 # plot traits by orders
   colors=wes_palette("Zissou1", 12,  type = "continuous")

   size_orders <- ggplot(traits_factors,aes(x=size, y=Order,color=Order)) + 
                    geom_point() + 
                    scale_color_manual(values=colors)+
                    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                          panel.background = element_blank(), axis.line = element_line(colour = "black"),
                          axis.text.y = element_text(color = colors))+ theme(legend.position="none")
    
   habitat_orders <- ggplot(traits_factors, aes(habitat)) + 
                      geom_bar(aes(fill = Order)) +
                      scale_fill_manual(values=colors)+
                      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                            panel.background = element_blank(), axis.line = element_line(colour = "black"),
                            legend.position="none",axis.text.x = element_text(angle = 45, hjust = 1))+
                      facet_wrap(vars(Order), scales = "free") 
    
   terrestriality_orders <- ggplot(traits_factors, aes(terrestriality)) + 
                              geom_bar(aes(fill = Order)) +
                              scale_fill_manual(values=colors)+
                              theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                                    panel.background = element_blank(), axis.line = element_line(colour = "black"),
                                    legend.position="none",axis.text.x = element_text(angle = 45, hjust = 1))+
                             facet_wrap(vars(Order), scales = "free") 
    
   vertical_orders <-  ggplot(traits_factors, aes(vertical)) + 
                        geom_bar(aes(fill = Order)) +
                        scale_fill_manual(values=colors)+
                        theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                              panel.background = element_blank(), axis.line = element_line(colour = "black"),
                              legend.position="none",axis.text.x = element_text(angle = 45, hjust = 1))+
                             facet_wrap(vars(Order), scales = "free") 
    
  diet_orders <- ggplot(traits_factors, aes(diet)) + 
                  geom_bar(aes(fill = Order)) +
                  scale_fill_manual(values=colors)+
                  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                        panel.background = element_blank(), axis.line = element_line(colour = "black"),
                        legend.position="none",axis.text.x = element_text(angle = 45, hjust = 1)) +
                  facet_wrap(vars(Order), scales = "free") 
    
  feeding_orders <- ggplot(traits_factors, aes(as.factor(feeding))) + 
                      geom_bar(aes(fill = Order)) +
                      scale_fill_manual(values=colors)+
                      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                      panel.background = element_blank(), axis.line = element_line(colour = "black"),
                      legend.position="none",axis.text.x = element_text(angle = 45, hjust = 1))  +
                      facet_wrap(vars(Order))+xlab("feeding")
    
  thermo_orders <- ggplot(traits_factors, aes(as.factor(thermo))) + 
                      geom_bar(aes(fill = Order)) +
                      scale_fill_manual(values=colors)+
                      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                      panel.background = element_blank(), axis.line = element_line(colour = "black"),
                      legend.position="none",axis.text.x = element_text(angle = 45, hjust = 1))+
                      facet_wrap(vars(Order))+xlab("thermo")
  
  iucn_orders <- ggplot(traits_factors, aes(as.factor(iucn_cat))) + 
                      geom_bar(aes(fill = Order)) +
                      scale_fill_manual(values=colors)+
                      theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                      panel.background = element_blank(), axis.line = element_line(colour = "black"),
                      legend.position="none",axis.text.x = element_text(angle = 45, hjust = 1))+
                      facet_wrap(vars(Order))+xlab("IUCN")
      
  pdf("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/traits.orders.pdf", 
  width = 20, height = 30)
  plot_grid(size_orders, habitat_orders, terrestriality_orders, vertical_orders, diet_orders, feeding_orders, 
             thermo_orders,  iucn_orders,
             labels= LETTERS[1:8], label_size = 12,align = "h", label_fontface = "bold", ncol = 2)
   dev.off()
  
  ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/traits.orders.jpg",
   plot = last_plot(),width = 30,height = 45, units = c("cm"), dpi = 300)
  
# test for correlation between traits
 model.matrix(~0+., data=traits_factors[, 2:8]) %>% 
     cor(use="pairwise.complete.obs", method = "spearman") %>% 
     ggcorrplot(show.diag=FALSE, type="lower", lab=TRUE, lab_size=2)

 traits_factors[, 2:8] %>%
   mutate_if(is.factor, as.numeric) %>% 
     cor(use="pairwise.complete.obs", method = "spearman") %>% 
     ggcorrplot(show.diag=FALSE, type="lower", lab=TRUE, lab_size=2)
 
 ggplot(traits_factors %>%
   mutate_if(is.factor, as.numeric) %>%
         pivot_longer(cols=habitat:size, names_to="trait", values_to="value"),
        aes(x=value, color = trait)) +
   geom_bar()+  facet_wrap(vars(trait), scales = "free")
 
```

# 4. PCoA

```{r chunk4, message=FALSE, warning=FALSE}
  library(clue)
  library(gtools)
  library(ape)
  library(geometry)
  library(ade4)

# How many dimensions better described our data?
# load function
  source('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Functions/quality_funct_space_fromdist.R', 
           chdir = TRUE)
# get dimensions
    quality=quality_funct_space_fromdist(distrait, nbdim=10, plot=NA)
    quality<- as.data.frame(quality$meanSD) # the best space is the one with the lowest mSD
    quality
    write.csv(quality, file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Tables/quality.csv")
  # extract number of dimensions
    which(quality$`quality$meanSD`==min(quality$`quality$meanSD`))
    plot(quality$`quality$meanSD`)
    order(quality$`quality$meanSD`)[1:10]

# Extract coordinates using PCoA
  pcoa<-dudi.pco(quasieuclid(distrait),scannf=F, nf=3)
# nf is the number of dimensions
  coords <- pcoa$li
  save(coords,file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/coords.RData')

# intertia explained
  sum(pcoa$eig[1:2])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:3])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:4])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:5])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:6])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:7])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:8])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:9])/sum(pcoa$eig)*100
  sum(pcoa$eig[1:10])/sum(pcoa$eig)*100
```

# 5. Calculate FD metrics per clade and status

```{r chunk5, message=FALSE,warning=FALSE}
 library(geometry)
 library(reshape2)

 source("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Functions/fonction_FRIC_Global_full.r") 
 source("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Functions/get_indicator_function 2.R")

# create presence and absence dataframes
  abun <- traits %>%
    dplyr::select(Species, Superorder) %>%
    pivot_wider(names_from=Superorder, values_from=Superorder) %>%
    left_join((shark.traits %>%
        dplyr::select(Species:Order.Torpediniformes)),
        by="Species") %>%
    dplyr::select(-size) %>%
    left_join((traits_imp %>%
                 dplyr::select(Species, iucn:southern))) %>%
    mutate(iucn_cat = recode(iucn, "0"="LC", "1"="NT", "2"="VU", "3"="EN","4"="CR")) %>%
    dplyr::select(-iucn) %>%
      pivot_wider(names_from=iucn_cat, values_from=iucn_cat)
  
  abun.superorder <- abun %>%
    dplyr::select(Species, Selachii, Batoidea) %>%
    mutate(Selachii=replace(Selachii, Selachii=="Selachii", 1)) %>%
    mutate(Batoidea=replace(Batoidea, Batoidea=="Batoidea", 1)) %>%
    column_to_rownames("Species") %>%
    replace(is.na(.), 0)
  #save(abun.superorder,file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/abun.superorder.RData')

  abun.order <- abun %>%
      dplyr::select(Species, Order.Carcharhiniformes:Order.Torpediniformes) %>%
      column_to_rownames("Species")

  abun.status <- abun %>%
      dplyr::select(Species, VU:CR) %>%
      column_to_rownames("Species") %>%
      mutate(across(where(is.factor), as.numeric)) %>%
      replace(!is.na(.), 1) %>%
      replace(is.na(.), 0)
  
  abun.status2 <- abun.status %>%
    mutate(AT=as.numeric(abun.status$VU)+as.numeric(abun.status$EN)+as.numeric(abun.status$CR)) %>%
    mutate(NoT=as.numeric(abun.status$LC)+as.numeric(abun.status$NT))%>%
    dplyr::select(AT, NoT)
  
  # abun.ocean <- abun %>%
  #   dplyr::select(Species, arctic:southern)%>%
  #   column_to_rownames("Species")
  
  # transform the dataframe in a list 
    test1 <- apply(t(abun.superorder),1,function(x){colnames(t(abun.superorder))[which(x==1)]}) 
    test2 <- apply(t(abun.order),1,function(x){colnames(t(abun.order))[which(x==1)]}) 
    test3 <- apply(t(abun.status),1,function(x){colnames(t(abun.status))[which(x==1)]}) 
    test3.2 <- apply(t(abun.status2),1,function(x){colnames(t(abun.status))[which(x==1)]}) 
    # test4 <- apply(t(abun.ocean),1,function(x){colnames(t(abun.ocean))[which(x==1)]}) 
    # test4$southern <- NULL # only 4 species here so not good.

  # apply to each element of the list the function get_FV_Sp that allows to calculate a lot of FD metrics. 
    lala1<-lapply(test1, function(y){ get_FV_Sp (ax=c(1:3), pcoa=pcoa, Selected_sp=y)})
    lala2<-lapply(test2, function(y){ get_FV_Sp (ax=c(1:3), pcoa=pcoa, Selected_sp=y)})
    lala3<-lapply(test3, function(y){ get_FV_Sp (ax=c(1:3), pcoa=pcoa, Selected_sp=y)})
    lala3.2<-lapply(test3.2, function(y){ get_FV_Sp (ax=c(1:3), pcoa=pcoa, Selected_sp=y)})
    # lala4<-lapply(test4, function(y){ get_FV_Sp (ax=c(1:3), pcoa=pcoa, Selected_sp=y)})
    
  # uniqueness
    distrait.uni<-as.matrix(distrait)
    lala1b<-lapply(test1, function(y){get_FU_region (Mat_dist=distrait.uni, nb_NN=5, Selected_sp=y)})
    lala2b<-lapply(test2, function(y){get_FU_region (Mat_dist=distrait.uni, nb_NN=5, Selected_sp=y)})
    lala3b<-lapply(test3, function(y){get_FU_region (Mat_dist=distrait.uni, nb_NN=5, Selected_sp=y)})
    lala3.2b<-lapply(test3.2, function(y){get_FU_region (Mat_dist=distrait.uni, nb_NN=5, Selected_sp=y)})

  # tables
    fric.superorder <- data.frame(
      sharks= rbind(round(lala1$Selachii$data$Fric_r, 3)*100,
                    lala1$Selachii$specialization[1],
                    lala1b$Selachii$FU_Region_Mean,
                    lala1$Selachii$data$RS),
      rays = rbind(round(lala1$Batoidea$data$Fric_r,3)*100,
                   lala1$Batoidea$specialization[1],
                   lala1b$Batoidea$FU_Region_Mean,
                   lala1$Batoidea$data$RS)
      )
    rownames(fric.superorder) <- c("fric", "fsp","fun", "s")
    colnames(fric.superorder) <- c("Selachii", "Batoidea")
    fric.superorder <- t(fric.superorder)

    fric.order <- data.frame(
      Carcharhiniformes= rbind(round(lala2$Order.Carcharhiniformes$data$Fric_r, 3)*100,
                               lala2$Order.Carcharhiniformes$specialization[1],
                               lala2b$Order.Carcharhiniformes$FU_Region_Mean,
                               lala2$Order.Carcharhiniformes$data$RS),
      Lamniformes = rbind(round(lala2$Order.Lamniformes$data$Fric_r,3)*100,
                          lala2$Order.Lamniformes$specialization[1],
                          lala2b$Order.Lamniformes$FU_Region_Mean,
                          lala2$Order.Lamniformes$data$RS),
      Orectolobiformes = rbind(round(lala2$Order.Orectolobiformes$data$Fric_r,3)*100,
                               lala2$Order.Orectolobiformes$specialization[1],
                               lala2b$Order.Orectolobiformes$FU_Region_Mean,
                               lala2$Order.Orectolobiformes$data$RS),
      Heterodontiformes = rbind(round(lala2$Order.Heterodontiformes$data$Fric_r,3)*100,
                                lala2$Order.Heterodontiformes$specialization[1],
                                lala2b$Order.Heterodontiformes$FU_Region_Mean,
                                lala2$Order.Heterodontiformes$data$RS),
      Squaliformes = rbind(round(lala2$Order.Squaliformes$data$Fric_r,3)*100,
                           lala2$Order.Squaliformes$specialization[1],
                           lala2b$Order.Squaliformes$FU_Region_Mean,
                           lala2$Order.Squaliformes$data$RS),
      Pristiophoriformes = rbind(round(lala2$Order.Pristiophoriformes$data$Fric_r,3)*100,
                                 lala2$Order.Pristiophoriformes$specialization[1],
                                 lala2b$Order.Pristiophoriformes$FU_Region_Mean,
                                 lala2$Order.Pristiophoriformes$data$RS),
      Squatiniformes = rbind(round(lala2$Order.Squatiniformes$data$Fric_r,3)*100,
                             lala2$Order.Squatiniformes$specialization[1],
                             lala2b$Order.Squatiniformes$FU_Region_Mean,
                             lala2$Order.Squatiniformes$data$RS),
      Hexanchiformes = rbind(round(lala2$Order.Hexanchiformes$data$Fric_r,3)*100,
                             lala2$Order.Hexanchiformes$specialization[1],
                             lala2b$Order.Hexanchiformes$FU_Region_Mean,
                             lala2$Order.Hexanchiformes$data$RS),
      Myliobatiformes = rbind(round(lala2$Order.Myliobatiformes$data$Fric_r,3)*100,
                              lala2$Order.Myliobatiformes$specialization[1],
                              lala2b$Order.Myliobatiformes$FU_Region_Mean,
                              lala2$Order.Myliobatiforme$data$RS),
      Rhinopristiformes = rbind(round(lala2$Order.Rhinopristiformes$data$Fric_r,3)*100,
                                lala2$Order.Rhinopristiformes$specialization[1],
                                lala2b$Order.Rhinopristiformes$FU_Region_Mean,
                                lala2$Order.Rhinopristiformes$data$RS),
      Torpediniformes = rbind(round(lala2$Order.Torpediniformes$data$Fric_r,3)*100,
                              lala2$Order.Torpediniformes$specialization[1],
                              lala2b$Order.Torpediniformes$FU_Region_Mean,
                              lala2$Order.Torpediniformes$data$RS),
      Rajiformes = rbind(round(lala2$Order.Rajiformes$data$Fric_r,3)*100,
                        lala2$Order.Rajiformes$specialization[1],
                        lala2b$Order.Rajiformes$FU_Region_Mean,
                        lala2$Order.Rajiformes$data$RS)
    )
    rownames(fric.order) <- c("fric", "fsp","fun", "s")
    colnames(fric.order) <- c("Carcharhiniformes", "Lamniformes", "Orectolobiformes", "Heterodontiformes", "Squaliformes", "Pristiophoriformes",
                              "Squatiniformes", "Hexanchiformes", "Myliobatiformes", "Rhinopristiformes", "Torpediniformes", "Rajiformes")
    fric.order <- t(fric.order)
    fric.order <- rbind(fric.superorder, fric.order)
    fric.order
    write.csv(fric.order, file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Tables/FD.order.csv")

    fric.iucn <- data.frame(
      #NE= round(lala3$NE$data$Fric_r, 3)*100,
      #DD = round(lala3$DD$data$Fric_r,3)*100,
      LC = rbind(round(lala3$LC$data$Fric_r,3)*100,
                lala3$LC$specialization[1],
                lala3b$LC$FU_Region_Mean,
                lala3$LC$data$RS),
      NT = rbind(round(lala3$NT$data$Fric_r,3)*100,
                 lala3$NT$specialization[1],
                 lala3b$NT$FU_Region_Mean,
                 lala3$NT$data$RS),
      VU = rbind(round(lala3$VU$data$Fric_r,3)*100,
                 lala3$VU$specialization[1],
                 lala3b$VU$FU_Region_Mean,
                 lala3$VU$data$RS),
      EN = rbind(round(lala3$EN$data$Fric_r,3)*100,
                 lala3$EN$specialization[1],
                 lala3b$EN$FU_Region_Mean,
                 lala3$EN$data$RS),
      CR = rbind(round(lala3$CR$data$Fric_r,3)*100,
                 lala3$CR$specialization[1],
                 lala3b$CR$FU_Region_Mean,
                 lala3$CR$data$RS),
      AT = rbind(round(lala3.2$AT$data$Fric_r,3)*100,
                 lala3.2$AT$specialization[1],
                 lala3.2b$AT$FU_Region_Mean,
                 lala3.2$AT$data$RS),
      NoT = rbind(round(lala3.2$NoT$data$Fric_r,3)*100,
                  lala3.2$NoT$specialization[1],
                 lala3.2b$NoT$FU_Region_Mean,
                 lala3.2$NoT$data$RS)
    )
    rownames(fric.iucn) <- c("fric", "fsp","fun", "s")
    colnames(fric.iucn) <- c("LC", "NT","VU", "EN", "CR", "AT", "NoT")
    fric.iucn <- t(fric.iucn)
    fric.iucn
    write.csv(fric.iucn, file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Tables/FD.iucn.csv")
    
    # fric.ocean <- data.frame(
    #   arctic = rbind(round(lala4$arctic$data$Fric_r,3)*100,
    #                        lala4$arctic$data$RS),
    #   north.atlantic = rbind(round(lala4$north.atlantic$data$Fric_r,3)*100,
    #                          lala4$north.atlantic$data$RS),      
    #   north.pacific = rbind(round(lala4$north.pacific$data$Fric_r,3)*100,
    #                               lala4$north.pacific$data$RS),
    #   indian = rbind(round(lala4$indian$data$Fric_r,3)*100,
    #                        lala4$indian$data$RS),
    #   south.pacific = rbind(round(lala4$south.pacific$data$Fric_r,3)*100,
    #                               lala4$south.pacific$data$RS),
    #   south.atlantic = rbind(round(lala4$south.atlantic$data$Fric_r,3)*100,
    #                                lala4$south.atlantic$data$RS)
    # )
    # rownames(fric.ocean) <- c("fric", "s")
    # fric.ocean 
  
```

# 6. Plot functional spaces
```{r chunk6, message=FALSE,warning=FALSE}
library(wesanderson)
library(RColorBrewer)
library(cowplot)
library(grid)
library (gghighlight)
library(visreg)
 options(ggrepel.max.overlaps = Inf)  
 
# find vertices of the 3D volume
   A1.max <- coords %>%
     rownames_to_column("Species") %>% 
     arrange(desc(A1)) %>% 
     slice(1:2) %>% 
     dplyr::select(Species)
   
   A1.min <- coords %>%
     rownames_to_column("Species") %>% 
     arrange(A1) %>% 
     slice(1:2) %>% 
     dplyr::select(Species)
   
    A2.max <- coords %>%
     rownames_to_column("Species") %>% 
     arrange(desc(A2)) %>% 
     slice(1:2) %>% 
     dplyr::select(Species)
   
   A2.min <- coords %>%
     rownames_to_column("Species") %>% 
     arrange(A2) %>% 
     slice(1:2) %>% 
     dplyr::select(Species)
   
   A3.max <- coords %>%
     rownames_to_column("Species") %>% 
     arrange(desc(A3)) %>% 
     slice(1:2) %>% 
     dplyr::select(Species)
   
   A3.min <- coords %>%
     rownames_to_column("Species") %>% 
     arrange(A3) %>% 
     slice(1:3) %>% 
     dplyr::select(Species)
 
 coords.groups <- coords %>%
      rownames_to_column("Species") %>%
      left_join(traits_factors, by="Species") %>%
      mutate(iucn_binary= recode(iucn_cat2, 
                        "NoT" = 0, 
                        "AT" = 1)) %>%
      mutate(vertice1=0) %>%
      mutate(vertice1 = replace(vertice1, Species %in% A1.max$Species, 1))%>%
      mutate(vertice1 = replace(vertice1, Species %in% A1.min$Species, 2))%>%
      
      mutate(vertice2=0)%>%
      mutate(vertice2 = replace(vertice2, Species %in% A2.max$Species, 1))%>%
      mutate(vertice2 = replace(vertice2, Species %in% A2.min$Species, 2))%>%
      
      mutate(vertice3=0)%>%
      mutate(vertice3 = replace(vertice3, Species %in% A3.max$Species, 1))%>%
      mutate(vertice3 = replace(vertice3, Species %in% A3.min$Species, 2))

# plots superorders
  colors=wes_palette("Zissou1", 12,  type = "continuous")
  levels(as.factor(unique(coords.groups$Order))) # order of colors for Orders
  
  # convex hulls
    hull.vol1 <- coords.groups %>%
      slice(chull(A1, A2))
    
    hull.vol2 <- coords.groups %>%
      slice(chull(A1, A3))
    
    hull.vol3 <- coords.groups %>%
      slice(chull(A2, A3))
    
    hull_superorders.vol1 <- coords.groups %>%
      group_by(Superorder) %>%
      slice(chull(A1, A2))
        
    hull_superorders.vol2 <- coords.groups %>%
      group_by(Superorder) %>%
      slice(chull(A1, A3))
     
    hull_superorders.vol3 <- coords.groups %>%
      group_by(Superorder) %>%
      slice(chull(A2, A3))  
    
    vol1.superorders <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = Superorder))+
      geom_point()+
      scale_color_manual(values=c(colors[1], colors[6]))+
      geom_text(data= coords.groups %>% filter(vertice1>0|vertice2>0),aes(label=Species), fontface = "italic")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Superorder)+
      geom_polygon(data=hull_superorders.vol1, alpha = 0.5)+
      scale_fill_manual(values=c(colors[1], colors[6])) +
      theme_bw()+
      theme_classic()+
      theme( legend.justification = c("right", "top"), 
              legend.box.background =element_blank())

   vol2.superorders <- 
        ggplot(coords.groups,aes(x=A1,y=A3, color = Superorder))+
        geom_point()+
        scale_color_manual(values=c(colors[1], colors[6]))+
        geom_text(data= coords.groups %>% filter(vertice1>0|vertice3>0),aes(label=Species), fontface = "italic") +
        geom_polygon(data=hull.vol2, fill=NA, color="black")+
        aes(fill=Superorder)+
        geom_polygon(data=hull_superorders.vol2, alpha = 0.5)+
        scale_fill_manual(values=c(colors[1], colors[6]))+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "none")
     
    vol3.superorders <- 
        ggplot(coords.groups,aes(x=A2,y=A3, color = Superorder))+
        geom_point()+
        scale_color_manual(values=c(colors[1], colors[6]))+
        geom_text(data= coords.groups %>% filter(vertice3>0|vertice3>0),aes(label=Species), fontface = "italic")+
        geom_polygon(data=hull.vol3, fill=NA, color="black")+
        aes(fill=Superorder)+
        geom_polygon(data=hull_superorders.vol3, alpha = 0.5)+
        scale_fill_manual(values=c(colors[1], colors[6]))+
        theme_bw()+         
        theme_classic()+
        theme(legend.position = "none")

    pdf("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/volumes.superorders.pdf", 
    width = 20,height = 8)
    
    plot_grid(vol1.superorders, vol2.superorders, vol3.superorders, labels= c("A","B", "C"),
                label_size = 12,align = "h", nrow=1, label_fontface = "bold")
    
   dev.off()

  # save
    ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/volumes.superorders.jpg",
    plot = last_plot(),width = 30,height = 10, units = c("cm"), dpi = 300)
    
    
# plot superorders separate 
 dark2_palette <- brewer.pal(8, "Dark2")
 colors=colorRampPalette(dark2_palette)(14)
 
  vol1.sharks <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Superorder)+
      geom_polygon(data=subset(hull_superorders.vol1, Superorder=="Selachii"),alpha = 0.5)+
      scale_fill_manual(values=colors[13:14])+
      geom_point(data=subset(coords.groups, Superorder=="Selachii"), color=colors[14])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Selachii")
  
   vol1.batoids <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Superorder)+
      geom_polygon(data=subset(hull_superorders.vol1, Superorder=="Batoidea"),alpha = 0.5)+
      scale_fill_manual(values=colors[13:14])+
      geom_point(data=subset(coords.groups, Superorder=="Batoidea"), color=colors[13])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Batoidea")    
    
# plots orders
   hull_orders.vol1 <- coords.groups %>%
      group_by(Order) %>%
      slice(chull(A1, A2))
 
 vol1.Carcharhiniformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Carcharhiniformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Carcharhiniformes"), color = colors[1])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Carcharhiniformes")
    
  vol1.Lamniformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Lamniformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Lamniformes"), color= colors[4])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Lamniformes")
  
    vol1.Orectolobiformes <- 
     ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Orectolobiformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Orectolobiformes"),color= colors[6])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Orectolobiformes")
    
    vol1.Heterodontiformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Heterodontiformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Heterodontiformes"), color=colors[2])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Heterodontiformes")
    
    vol1.Squaliformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Squaliformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Squaliformes"),color=colors[10])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Squaliformes")
    
    vol1.Pristiophoriformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Pristiophoriformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Pristiophoriformes"), color=colors[7])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Pristiophoriformes")
    
    vol1.Squatiniformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Squatiniformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Squatiniformes"), color=colors[11])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Squatiniformes")
    
     vol1.Hexanchiformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Hexanchiformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Hexanchiformes"), color=colors[3])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Hexanchiformes")
     
    vol1.Myliobatiformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Myliobatiformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Myliobatiformes"), color=colors[5])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Myliobatiformes")
    
    vol1.Rhinopristiformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Rhinopristiformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Rhinopristiformes"),color=colors[9])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Rhinopristiformes")
    
    vol1.Torpediniformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Torpediniformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Torpediniformes"),color=colors[12])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Torpediniformes")
    
    vol1.Rajiformes <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = "grey"))+
      geom_point()+
      scale_color_manual(values="grey")+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=Order)+
      geom_polygon(data=subset(hull_orders.vol1, Order=="Rajiformes"),alpha = 0.5)+
      scale_fill_manual(values=colors)+
      geom_point(data=subset(coords.groups, Order=="Rajiformes"), color=colors[8])+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")+
      labs(title="Rajiformes")
    
    plot_grid(vol1.Carcharhiniformes, vol1.Lamniformes, vol1.Orectolobiformes, 
            vol1.Heterodontiformes, vol1.Squaliformes, vol1.Pristiophoriformes,
            vol1.Squatiniformes, vol1.Hexanchiformes, vol1.Myliobatiformes, 
            vol1.Rhinopristiformes, vol1.Torpediniformes, vol1.Rajiformes,  nrow=3)
  # save 
    ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/volumes.orders.jpg",
    plot = last_plot(),width = 30,height = 15, units = c("cm"), dpi = 300)
  
# plot IUCN volumes
  
  hull_iucn.vol1 <- coords.groups %>%
      group_by(iucn_cat) %>%
      slice(chull(A1, A2))
  
  iucn.cols=c(
         "LC" = "forestgreen", 
         "NT" = "yellowgreen",
         "VU" = "gold", 
         "EN" = "orange", 
         "CR" = "red")
  
  iucn.cols2=c(
         "NoT" = "forestgreen", 
         "AT" = "red")
  
    vol1.LC <- 
        ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat))+
        geom_point()+
        scale_color_manual(values=iucn.cols)+
        geom_polygon(data=hull.vol1, fill=NA, color="black")+
        aes(fill=iucn_cat)+
        geom_polygon(data=subset(hull_iucn.vol1, iucn_cat=="LC"),alpha=0.5)+
        scale_fill_manual(values=iucn.cols)+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "LC")+
        labs(title="LC")
    
   vol1.NT <- 
        ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat))+
        geom_point()+
        scale_color_manual(values=iucn.cols)+
        geom_polygon(data=hull.vol1, fill=NA, color="black")+
        aes(fill=iucn_cat)+
        geom_polygon(data=subset(hull_iucn.vol1, iucn_cat=="NT"),alpha=0.5)+
        scale_fill_manual(values=iucn.cols)+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "NT")+
        labs(title="NT")
   
   vol1.VU <- 
        ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat))+
        geom_point()+
        scale_color_manual(values=iucn.cols)+
        geom_polygon(data=hull.vol1, fill=NA, color="black")+
        aes(fill=iucn_cat)+
        geom_polygon(data=subset(hull_iucn.vol1, iucn_cat=="VU"),alpha=0.5)+
        scale_fill_manual(values=iucn.cols)+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "VU")+
        labs(title="VU")
   
   vol1.EN <- 
        ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat))+
        geom_point()+
        scale_color_manual(values=iucn.cols)+
        geom_polygon(data=hull.vol1, fill=NA, color="black")+
        aes(fill=iucn_cat)+
        geom_polygon(data=subset(hull_iucn.vol1, iucn_cat=="EN"),alpha=0.5)+
        scale_fill_manual(values=iucn.cols)+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "EN")+
        labs(title="EN")
   
     vol1.CR <- 
        ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat))+
        geom_point()+
        scale_color_manual(values=iucn.cols)+
        geom_polygon(data=hull.vol1, fill=NA, color="black")+
        aes(fill=iucn_cat)+
        geom_polygon(data=subset(hull_iucn.vol1, iucn_cat=="CR"),alpha=0.5)+
        scale_fill_manual(values=iucn.cols)+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "CR")+
        labs(title="CR")
    
  hull_iucn2.vol1 <- coords.groups %>%
      group_by(iucn_cat2) %>%
      slice(chull(A1, A2))
  
  hull_iucn2.vol2 <- coords.groups %>%
      group_by(iucn_cat2) %>%
      slice(chull(A1, A3))
  
  vol1.no.threat <- 
        ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat2))+
        geom_point()+
        scale_color_manual(values=iucn.cols2)+
        geom_polygon(data=hull.vol1, fill=NA, color="black")+
        aes(fill=iucn_cat2)+
        geom_polygon(data=subset(hull_iucn2.vol1, iucn_cat2=="NoT"),alpha=0.5)+
        scale_fill_manual(values=iucn.cols2)+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "Not threatened")+
        labs(title="Not threatened")
  
  vol1.threatened <- 
        ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat2))+
        geom_point()+
        scale_color_manual(values=iucn.cols2)+
        geom_polygon(data=hull.vol1, fill=NA, color="black")+
        aes(fill=iucn_cat2)+
        geom_polygon(data=subset(hull_iucn2.vol1, iucn_cat2=="AT"),alpha=0.5)+
        scale_fill_manual(values=iucn.cols2)+
        theme_bw()+
        theme_classic()+
        theme(legend.position = "Threatened")+
        labs(title="Threatened")
  
  vol2.threat <- 
      ggplot(coords.groups,aes(x=A1,y=A3, color = iucn_cat2))+
      geom_point()+
      scale_color_manual(values=iucn.cols2)+
      geom_text(data= coords.groups %>% filter(vertice1>0|vertice3>0),aes(label=Species))+
      geom_polygon(data=hull.vol2, fill=NA, color="black")+
      aes(fill=iucn_cat2)+
      geom_polygon(data=hull_iucn2.vol2, alpha = 0.5)+
      scale_fill_manual(values=c(iucn.cols2))+
      theme_bw()+
      theme_classic()

    vol1.threat_paper <- 
      ggplot(coords.groups,aes(x=A1,y=A2, color = iucn_cat2))+
      geom_point()+
      scale_color_manual(values=iucn.cols2)+
      #geom_text(data= coords.groups %>% filter(vertice1>0|vertice3>0),aes(label=Species))+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      aes(fill=iucn_cat2)+
      geom_polygon(data=hull_iucn2.vol1, alpha = 0.5)+
      scale_fill_manual(values=c(iucn.cols2))+
      theme_bw()+
      theme_classic()+
      labs(title="Extinction risk")+
      theme(legend.position = "none")
  
  plot_grid(vol1.LC,vol1.NT, vol1.VU, vol1.EN,vol1.CR,
            labels= LETTERS[1:8], label_size = 12,align = "h", label_fontface = "bold",  ncol=2)
      
  # save 
    ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/volumes.iucn.jpg",
    plot = last_plot(),width = 15,height = 20, units = c("cm"), dpi = 300)
    

    plot_grid(vol1.sharks, vol1.Carcharhiniformes, vol1.Lamniformes, vol1.Orectolobiformes, vol1.Heterodontiformes,
              vol1.batoids,vol1.Squaliformes, vol1.Pristiophoriformes,vol1.Squatiniformes, vol1.Hexanchiformes,
              vol1.threat_paper,vol1.Myliobatiformes,vol1.Rhinopristiformes, vol1.Torpediniformes, vol1.Rajiformes,
              nrow=3)
  
  ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/Fig1A.jpg",
  plot = last_plot(),width = 30,height = 12, units = c("cm"), dpi = 300)
       
# # Biomes
#   hull_oceans.vol1 <- coords.groups %>%
#     dplyr::select(-Order, -Superorder, -iucn_cat) %>%
#     pivot_longer(cols=arctic:southern,
#                  names_to="ocean", values_to="value") %>%
#     filter(value==1) %>%
#     group_by(ocean) %>%
#     slice(chull(A1, A2))
#   
#   ocean.cols= c('#d53e4f','#fdae61','#fee08b','#e6f598','#abdda4','#66c2a5',"black")
# 
#   vol1.arctic <- 
#         ggplot(hull_oceans.vol1,aes(x=A1,y=A2, color = ocean))+
#         geom_point()+
#         scale_color_manual(values=ocean.cols)+
#         geom_polygon(data=hull.vol1, fill=NA, color="black")+
#         aes(fill=ocean)+
#         geom_polygon(data=subset(hull_oceans.vol1, ocean=="arctic"),alpha=0.5)+
#         scale_fill_manual(values=ocean.cols)+
#         theme_bw()+
#         theme_classic()+
#         theme(legend.position = "arctic")+
#         labs(title="arctic")
#     
#   vol1.north.atlantic <- 
#         ggplot(hull_oceans.vol1,aes(x=A1,y=A2, color = ocean))+
#         geom_point()+
#         scale_color_manual(values=ocean.cols)+
#         geom_polygon(data=hull.vol1, fill=NA, color="black")+
#         aes(fill=ocean)+
#         geom_polygon(data=subset(hull_oceans.vol1, ocean=="north.atlantic"),alpha=0.5)+
#         scale_fill_manual(values=ocean.cols)+
#         theme_bw()+
#         theme_classic()+
#         theme(legend.position = "north.atlantic")+
#         labs(title="north.atlantic")
#   
#   vol1.north.pacific <- 
#         ggplot(hull_oceans.vol1,aes(x=A1,y=A2, color = ocean))+
#         geom_point()+
#         scale_color_manual(values=ocean.cols)+
#         geom_polygon(data=hull.vol1, fill=NA, color="black")+
#         aes(fill=ocean)+
#         geom_polygon(data=subset(hull_oceans.vol1, ocean=="north.pacific"),alpha=0.5)+
#         scale_fill_manual(values=ocean.cols)+
#         theme_bw()+
#         theme_classic()+
#         theme(legend.position = "north.pacific")+
#         labs(title="north.pacific")
#   
#     vol1.indian <- 
#         ggplot(hull_oceans.vol1,aes(x=A1,y=A2, color = ocean))+
#         geom_point()+
#         scale_color_manual(values=ocean.cols)+
#         geom_polygon(data=hull.vol1, fill=NA, color="black")+
#         aes(fill=ocean)+
#         geom_polygon(data=subset(hull_oceans.vol1, ocean=="indian"),alpha=0.5)+
#         scale_fill_manual(values=ocean.cols)+
#         theme_bw()+
#         theme_classic()+
#         theme(legend.position = "indian")+
#         labs(title="indian")
#     
#     vol1.south.pacific <- 
#         ggplot(hull_oceans.vol1,aes(x=A1,y=A2, color = ocean))+
#         geom_point()+
#         scale_color_manual(values=ocean.cols)+
#         geom_polygon(data=hull.vol1, fill=NA, color="black")+
#         aes(fill=ocean)+
#         geom_polygon(data=subset(hull_oceans.vol1, ocean=="south.pacific"),alpha=0.5)+
#         scale_fill_manual(values=ocean.cols)+
#         theme_bw()+
#         theme_classic()+
#         theme(legend.position = "south.pacific")+
#         labs(title="south.pacific")
#     
#     vol1.south.atlantic <- 
#         ggplot(hull_oceans.vol1,aes(x=A1,y=A2, color = ocean))+
#         geom_point()+
#         scale_color_manual(values=ocean.cols)+
#         geom_polygon(data=hull.vol1, fill=NA, color="black")+
#         aes(fill=ocean)+
#         geom_polygon(data=subset(hull_oceans.vol1, ocean=="south.atlantic"),alpha=0.5)+
#         scale_fill_manual(values=ocean.cols)+
#         theme_bw()+
#         theme_classic()+
#         theme(legend.position = "south.atlantic")+
#         labs(title="south.atlantic")
# 
#   plot_grid(vol1.arctic, vol1.north.atlantic,vol1.north.pacific, 
#             vol1.indian, vol1.south.pacific, vol1.south.atlantic,
#             labels= letters[1:6], label_size = 12,align = "h", label_fontface = "bold",  ncol=2)
#       
#   # save 
#     ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Plots/volumes.oceans.jpg",
#     plot = last_plot(),width = 15,height = 20, units = c("cm"), dpi = 300)
#     
#     pdf("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Plots/volumes.oceans.pdf", 
#         width = 8, height = 10)
#         plot_grid(vol1.arctic, vol1.north.atlantic,vol1.north.pacific, 
#             vol1.indian, vol1.south.pacific, vol1.south.atlantic, 
#             labels= letters[1:6], label_size = 12,align = "h", label_fontface = "bold",  ncol=2)
#         dev.off()
  
```
# 7. Relationship between coordinates and traits and IUCN

```{r chunk7, message=FALSE,warning=FALSE}
library(robustbase)
# linear models of coords vs traits
# test for normality
shapiro.test(coords.groups$A1)
shapiro.test(coords.groups$A2)
shapiro.test(coords.groups$A3)

  # habitat
    habitatA1=lmrob(coords.groups$A1~coords.groups$habitat)
    habitatA2=lmrob(coords.groups$A2~coords.groups$habitat)
    habitatA3=lmrob(coords.groups$A3~coords.groups$habitat)
    habitat_axes<-data.frame(A1=summary(habitatA1)$r.squared,
                          A2=summary(habitatA2)$r.squared,
                          A3=summary(habitatA3)$r.squared)
    
    habitat_threat=glm(iucn_binary ~ habitat, family = "binomial", coords.groups)
    habitat_drop <-drop1(habitat_threat, test = "Chi")
    habitat_p<-c(habitat_drop$Df[2], habitat_drop$LRT[2], habitat_drop$`Pr(>Chi)`[2])
    
  # terrestriality
    terrestrialityA1=lmrob(coords.groups$A1~coords.groups$terrestriality)
    terrestrialityA2=lmrob(coords.groups$A2~coords.groups$terrestriality)
    terrestrialityA3=lmrob(coords.groups$A3~coords.groups$terrestriality)
    terrestriality_axes<-data.frame(A1=summary(terrestrialityA1)$r.squared,
                          A2=summary(terrestrialityA2)$r.squared,
                          A3=summary(terrestrialityA3)$r.squared)
    
    terrestriality_threat=glm(iucn_binary ~ terrestriality, family = "binomial", coords.groups)
    terrestriality_drop <-drop1(terrestriality_threat, test = "Chi")
    terrestriality_p<-c(terrestriality_drop$Df[2], terrestriality_drop$LRT[2], terrestriality_drop$`Pr(>Chi)`[2])
    
  # diet
    dietA1=lmrob(coords.groups$A1~coords.groups$diet)
    dietA2=lmrob(coords.groups$A2~coords.groups$diet)
    dietA3=lmrob(coords.groups$A3~coords.groups$diet)
    diet_axes<-data.frame(A1=summary(dietA1)$r.squared,
                          A2=summary(dietA2)$r.squared,
                          A3=summary(dietA3)$r.squared)  
    
    diet_threat=glm(iucn_binary ~ diet, family = "binomial", coords.groups)
    diet_drop <-drop1(diet_threat, test = "Chi")
    diet_p<-c(diet_drop$Df[2], diet_drop$LRT[2], diet_drop$`Pr(>Chi)`[2])
    
  # size
   sizeA1=lmrob(coords.groups$A1~coords.groups$size)
   sizeA2=lmrob(coords.groups$A2~coords.groups$size)
   sizeA3=lmrob(coords.groups$A3~coords.groups$size)
   size_axes<-data.frame(A1=summary(sizeA1)$r.squared,
                        A2=summary(sizeA2)$r.squared,
                        A3=summary(sizeA3)$r.squared)
   
   size_threat=glm(iucn_binary ~ size, family = "binomial", coords.groups)
   size_drop <-drop1(size_threat, test = "Chi")
   size_p<-c(size_drop$Df[2], size_drop$LRT[2], size_drop$`Pr(>Chi)`[2])
  
  # vertical
    verticalA1=lmrob(coords.groups$A1~coords.groups$vertical)
    verticalA2=lmrob(coords.groups$A2~coords.groups$vertical)
    verticalA3=lmrob(coords.groups$A3~coords.groups$vertical)
    vertical_axes<-data.frame(A1=summary(verticalA1)$r.squared,
                          A2=summary(verticalA2)$r.squared,
                          A3=summary(verticalA3)$r.squared) 
  
    vertical_threat=glm(iucn_binary ~ vertical, family = "binomial", coords.groups)
    vertical_drop <-drop1(vertical_threat, test = "Chi")
    vertical_p<-c(vertical_drop$Df[2], vertical_drop$LRT[2], vertical_drop$`Pr(>Chi)`[2])  
    
  # thermo
    thermoA1=lmrob(coords.groups$A1~coords.groups$thermo)
    thermoA2=lmrob(coords.groups$A2~coords.groups$thermo)
    thermoA3=lmrob(coords.groups$A3~coords.groups$thermo)
    thermo_axes<-data.frame(A1=summary(thermoA1)$r.squared,
                            A2=summary(thermoA2)$r.squared,
                            A3=summary(thermoA3)$r.squared)  
      
    thermo_threat=glm(iucn_binary ~ thermo, family = "binomial", coords.groups)
    thermo_drop <-drop1(thermo_threat, test = "Chi")
    thermo_p<-c(thermo_drop$Df[2], thermo_drop$LRT[2], thermo_drop$`Pr(>Chi)`[2]) 
    
  # feeding
    feedingA1=lmrob(coords.groups$A1~coords.groups$feeding)
    feedingA2=lmrob(coords.groups$A2~coords.groups$feeding)
    feedingA3=lmrob(coords.groups$A3~coords.groups$feeding)
    feeding_axes<-data.frame(A1=summary(feedingA1)$r.squared,
                            A2=summary(feedingA2)$r.squared,
                            A3=summary(feedingA3)$r.squared)  
    
    feeding_threat=glm(iucn_binary ~ feeding, family = "binomial", coords.groups)
    feeding_drop <-drop1(feeding_threat, test = "Chi")
    feeding_p<-c(feeding_drop$Df[2], feeding_drop$LRT[2], feeding_drop$`Pr(>Chi)`[2]) 
      
    axes_traits<-rbind(habitat_axes, terrestriality_axes,diet_axes, size_axes,vertical_axes,
                       thermo_axes,feeding_axes)
    rownames(axes_traits)<-c("habitat","terrestriality","diet","size", "vertical", "thermo", "feeding")
    which(axes_traits$A1==max(axes_traits$A1))
    which(axes_traits$A2==max(axes_traits$A2))
    which(axes_traits$A3==max(axes_traits$A3))
          
    axes_traits
    write.csv(axes_traits, file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Tables/axes_traits_lmrob.csv")

    aic <- AIC(habitat_threat,terrestriality_threat,diet_threat,size_threat,vertical_threat,thermo_threat,feeding_threat)
    
    threat_traits<-rbind(habitat_p,terrestriality_p,diet_p,size_p,vertical_p,thermo_p,feeding_p)
    colnames(threat_traits) <- c("DF", "LRT", "Pr(>Chi)")
    
    threat_traits
    write.csv(threat_traits, file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Tables/threat_traits.csv")

# Plot volumes of IUCN status   
  pdf("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/glms.iucn.pdf", 
        width = 10, height = 15)
  par(mfrow=c(4,2))  
  
  visreg(habitat_threat, "habitat", scale = "response", ylab = "threat")
  visreg(terrestriality_threat, "terrestriality", scale = "response", ylab = "threat")
  visreg(diet_threat, "diet", scale = "response", ylab = "threat")
  visreg(size_threat, "size", scale = "response", ylab = "threat")
  visreg(vertical_threat, "vertical", scale = "response", ylab = "threat")
  #visreg(thermo_threat, "thermo", scale = "response", ylab = "threat")
  visreg(feeding_threat, "feeding", scale = "response", ylab = "threat")
        
  dev.off()
```

# 8. Calculate FD metrics per species
```{r chunk8, message=FALSE,warning=FALSE}
library(stringr)
library(cowplot)
library(vegan)
library(RColorBrewer)
library(gghighlight)
library(mFD)
library(tidyverse)

# source("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Functions/iucn_palette.R")
# source('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Functions/FUSE_function_end.R', chdir = TRUE)
# load('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/distrait.RData')
# load('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/coords.RData')

# convert IUCN categories to ext prob
   traits_imp <- traits_imp %>%
    mutate(ext_prob100 = recode(iucn, "0"=0.001748844, "1"=0.014096969, "2"=0.1, "3"=0.67232,"4"=0.999023438))
   
  FUSE_res <- fuse(as.matrix(distrait), as.matrix(coords), nb_NN=5, GE=traits_imp$ext_prob100,standGE = FALSE) #mFD package

# FUSE_res<-FUSE(Mat_dist=distrait,Coords=coords,nb_NN=5,GE=traits_imp$ext_prob100,StandGE=F) #extracted function 
 
# dataset with scores and taxonomy
  FUSE_sp <- FUSE_res %>%
      dplyr::select(FUSE, FUn_std, FSp_std) %>%
      rename(FSp = FSp_std, FUn = FUn_std) %>%
      rownames_to_column("Species") %>%
      left_join(traits %>%
                dplyr::select(Species, Order, Superorder), by="Species")

  ED2 <- readRDS("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_inputs/Res_HED_sharks.rds") %>%
         mutate(names_from = stringr::str_replace(Species, "_", " ")) %>%
         dplyr::select(-Species, -HED_sd) %>%
         rename(Species=names_from, ED2=HED_mean)
  
  EDGE2 <- readRDS("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_inputs/Res_HEDGE_sharks.rds") %>%
         mutate(names_from = stringr::str_replace(Species, "_", " ")) %>%
         dplyr::select(-Species, -HEDGE_sd) %>%
         rename(Species=names_from, EDGE2=HEDGE_mean)

# combine FUSE and EDGE metrics
  shark_sp_metrics <- as_tibble(FUSE_sp)%>%
    left_join(traits_imp %>%
               dplyr::select(Species, iucn, ext_prob100), by="Species")%>%
    mutate(iucn_status = recode(iucn, "0"="LC", "1"="NT", "2"="VU", "3"="EN","4"="CR"))%>%
    left_join(ED2, by="Species") %>%
    left_join(EDGE2, by="Species") %>%
    left_join(coords.groups %>%
                dplyr:: select(Species, iucn_cat2), by= "Species")
    shark_sp_metrics$iucn_status <- as.factor(shark_sp_metrics$iucn_status)
    shark_sp_metrics$iucn_status <- fct_relevel(shark_sp_metrics$iucn_status, c("CR","EN","VU","NT","LC"))
    
  write.csv(shark_sp_metrics, 
            file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/shark_sp_metrics.csv")
    
# top species
  top_species <- shark_sp_metrics %>%
                slice_max(FUn, n = 20) %>%
                dplyr::select(Species, FUn, iucn_status) %>%
                rename(top_FUn=Species) %>%
                mutate(top_FSp=
                         shark_sp_metrics %>%
                         slice_max(FSp, n = 20) %>%
                         dplyr::select(Species, FSp, iucn_status))%>%
                mutate(top_FUSE=
                         shark_sp_metrics %>%
                         slice_max(FUSE, n = 20) %>%
                         dplyr::select(Species, FUSE, iucn_status)) %>%
                mutate(top_ED2=
                         shark_sp_metrics %>%
                         slice_max(ED2, n = 20) %>%
                         dplyr::select(Species, ED2, iucn_status)) %>%
                mutate(top_EDGE2=
                         shark_sp_metrics %>%
                         slice_max(EDGE2, n = 20) %>%
                         dplyr::select(Species, EDGE2, iucn_status))
  
# check associations between top species and their traits
  coords.groups <- coords.groups%>%
      mutate(top_fun=0)%>%
      mutate(top_fun=replace(top_fun, Species %in% top_species$top_FUn[1:3], 1:3))%>%
      
      mutate(top_fsp=0)%>%
      mutate(top_fsp=replace(top_fsp, Species %in% top_species$top_FSp$Species[1:3], 1:3))%>%
      
      mutate(top_ed2=0)%>%
      mutate(top_ed2=replace(top_ed2, Species %in% top_species$top_ED2$Species[1:5], 1:5))%>%
    
      mutate(top_fuse_edge=0)%>%
      mutate(top_fuse_edge=replace(top_fuse_edge, Species %in% top_species$top_FUSE$Species[1:20], "fuse"))%>%
      mutate(top_fuse_edge=replace(top_fuse_edge, Species %in% top_species$top_EDGE2$Species[1:20], "edge"))%>%
      mutate(top_fuse_edge = replace(top_fuse_edge, Species %in% intersect(
        top_species$top_EDGE2$Species, top_species$top_FUSE$Species),"both")) %>%
    
     mutate(top_edge2=0)%>%
      mutate(top_edge2=replace(top_edge2, Species %in% top_species$top_EDGE2$Species[1:20], 1:20))%>%
    
     mutate(top_fuse=0)%>%
      mutate(top_fuse=replace(top_fuse, Species %in% top_species$top_FUSE$Species[1:20], 1:20))

  top_traits <- coords.groups%>%
      subset(top_fun>0|top_fsp>0|top_fuse_edge!="0")%>%
      dplyr::select(c(Species, habitat:size))

 top_traits

# plot values per order
   colors=colorRampPalette(dark2_palette)(14)

  orders_species_metrics <-  shark_sp_metrics %>% 
        group_by(Order)%>% 
        summarise(Species= n_distinct(Species),
                  FSp = median(FSp),
                  FUn = median(FUn), 
                  FUSE = median(FUSE),
                  ED2 = median(ED2),
                  EDGE = median (EDGE2))
  
  orders_species_metrics
  write.csv(orders_species_metrics, 
            file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Tables/orders_species_metrics.csv")
    
  iucn_species_metrics <- shark_sp_metrics %>% 
        group_by(iucn_cat2)%>% 
            summarise(Species= n_distinct(Species),
                      FSp = mean(FSp),
                      FUn = mean(FUn),
                      ED2 = mean (ED2))
  
  iucn_species_metrics 
  
  write.csv(iucn_species_metrics, 
            file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Tables/iucn_species_metrics.csv")
  
  FUn_orders <- ggplot(shark_sp_metrics,aes(y=reorder(Order, FUn, median),x=log(FUn)))+
                geom_boxplot(aes(fill=Order))+
                scale_fill_manual(values=colors) +
                theme_classic()+guides(fill="none")+
                labs(x = "log(FUn)")+theme(axis.title.y=element_blank())
   
  FSp_orders <-  ggplot(shark_sp_metrics,aes(y=reorder(Order, FSp, median),x=FSp))+
                 geom_boxplot(aes(fill=Order))+
                 scale_fill_manual(values=colors) +
                 theme_classic()+guides(fill="none")+
                 labs(x = "FSp")+theme(axis.title.y=element_blank())
 
  FUSE_orders <- ggplot(shark_sp_metrics,aes(y=reorder(Order, FUSE, median),x=log(FUSE)))+
                 geom_boxplot(aes(fill=Order))+
                 scale_fill_manual(values=colors)+
                 theme_classic()+guides(fill="none")+
                 labs(x = "log(FUSE)")+theme(axis.title.y=element_blank())
  
  IUCN_orders <- ggplot(shark_sp_metrics,aes(y=reorder(Order, ext_prob100, median),x=ext_prob100))+
                 geom_boxplot(aes(fill=Order))+
                 scale_fill_manual(values=colors)+
                 theme_classic()+guides(fill="none")+
                 labs(x = "IUCN (p_100)")+theme(axis.title.y=element_blank())
   
  ED2_FD_IUCN <- shark_sp_metrics %>%
    dplyr::select(Species, FSp, FUn, ED2, FUSE, EDGE2, iucn_cat2, Order) %>%
    pivot_longer(cols=FSp:ED2, names_to="metric", values_to="values")%>%
    group_by(iucn_cat2)
    
    
  ED2_FD_IUCN_plot <-  ggplot(ED2_FD_IUCN,aes(x=iucn_cat2,y=log(values)))+
                          geom_boxplot(aes(fill=iucn_cat2))+
                          facet_wrap(vars(metric), scales = "free")+
                          scale_fill_manual(values=c("red","forestgreen"))+
                          guides(fill="none")+
                          theme(axis.title.y=element_blank())+xlab("Level of threat")
  
# test normality to compare metrics  
  shapiro.test(shark_sp_metrics$FUn)  
  shapiro.test(shark_sp_metrics$FSp)
  shapiro.test(shark_sp_metrics$ED2)
  
   ED2_AT <-  shark_sp_metrics %>%
     dplyr::select(ED2, iucn_cat2) %>%
     filter(iucn_cat2=="AT")%>%
      dplyr::select(ED2)
       
    ED2_NoT <- shark_sp_metrics %>%
      dplyr::select(ED2, iucn_cat2) %>%
      filter(iucn_cat2=="NoT") %>%
      dplyr::select(ED2)
    
   ED2_test <-wilcox.test(ED2_AT$ED2, ED2_NoT$ED2)
   
   FSp_AT <-  shark_sp_metrics %>%
        dplyr::select(FSp, iucn_cat2) %>%
        filter(iucn_cat2=="AT")%>%
        dplyr::select(FSp)
       
    FSp_NoT <- shark_sp_metrics %>%
        dplyr::select(FSp, iucn_cat2) %>%
        filter(iucn_cat2=="NoT") %>%
        dplyr::select(FSp)
      
   FSp_test <- wilcox.test(FSp_AT$FSp, FSp_NoT$FSp)

    FUn_AT <- shark_sp_metrics %>%
        dplyr::select(FUn, iucn_cat2) %>%
        filter(iucn_cat2=="AT")%>%
        dplyr::select(FUn)
       
     FUn_NoT <- shark_sp_metrics %>%
          dplyr::select(FUn, iucn_cat2) %>%
          filter(iucn_cat2=="NoT") %>%
          dplyr::select(FUn)
     
  FUn_test <- wilcox.test(FUn_AT$FUn, FUn_NoT$FUn)

  ED2_test$p.value 
  FSp_test$p.value
  FUn_test$p.value  

 plot_grid(FUn_orders, FSp_orders, FUSE_orders, nrow=1)

  ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/Orders_mean_values.jpg",
  plot = last_plot(),width = 30,height = 20, units = c("cm"), dpi = 300)
  
  plot_grid(IUCN_orders, ED2_FD_IUCN_plot , nrow=1, labels= c("A","B", "C"),
                label_size = 12,align = "h", label_fontface = "bold")
  ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/iucn_ed_fd.jpg",
  plot = last_plot(),width = 25,height = 10, units = c("cm"), dpi = 300)


  vol1.top_fuse_edge <-
      ggplot(coords.groups,aes(x=A1,y=A2, color = top_fuse_edge))+
      geom_point(size=4)+
      #geom_point(data=subset(coords.groups, top_fuse_edge!="0"))+
      gghighlight(top_fuse_edge!="0", label_key = Species) +
      # gghighlight(top_fuse_edge=="fuse"|top_fuse_edge=="edge", label_key = Species) +
      # geom_point(data=subset(coords.groups, top_fuse_edge=="both"),
      #       aes(x=A1,y=A2, color = "#C77EFF"))+
      # geom_text(data=subset(coords.groups, top_fuse_edge!="0"),
      #       aes(label=Species))+
      geom_polygon(data=hull.vol1, fill=NA, color="black")+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")

    vol2.top_fuse_edge <-
      ggplot(coords.groups,aes(x=A1,y=A3, color = top_fuse_edge))+
      geom_point(size=4)+
      gghighlight(top_fuse_edge!="0", label_key = Species) +
      geom_polygon(data=hull.vol2, fill=NA, color="black")+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")

    vol3.top_fuse_edge <-
      ggplot(coords.groups,aes(x=A2,y=A3, color = top_fuse_edge))+
      geom_point(size=4)+
      gghighlight(top_fuse_edge!="0", label_key = Species) +
      geom_polygon(data=hull.vol3, fill=NA, color="black")+
      theme_bw()+
      theme_classic()+
      theme(legend.position = "none")

  plot_grid(vol1.top_fuse_edge, vol2.top_fuse_edge, vol3.top_fuse_edge,
                label_size = 12,align = "h", nrow=1, label_fontface = "bold")

  # save
    ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/volumes.fuse_edge.jpg",
    plot = last_plot(),width = 40,height = 20, units = c("cm"), dpi = 300)

  
shapiro.test(shark_sp_metrics$FUn)
shapiro.test(shark_sp_metrics$FSp)
shapiro.test(shark_sp_metrics$ED2)
  
  cor.test(shark_sp_metrics$ED2, shark_sp_metrics$FUn,
         method = "spearman")
  
  FUn_ED <- ggplot(shark_sp_metrics,aes(y=FUn,x=ED2))+
              geom_point(aes(col=ext_prob100))+
              scale_colour_gradientn(colours = c("forestgreen", "orange", "red"))+
              theme_classic()+ylab("FUn")+
            annotate(geom="text", 180, y = 0.96, label="rho=0.22")
  
  cor.test(shark_sp_metrics$ED2, shark_sp_metrics$FSp,
         method = "spearman")
  
  # qplot(ED2,FUn, facets = ~ Order, data = shark_sp_metrics)+
  #   geom_smooth(method = lm, se=F)
  # 
  #  qplot(ED,FUn, facets = ~ iucn, data = shark_sp_metrics)+
  #   geom_smooth(method = lm, se=F)
  
  # glm(shark_sp_metrics$ED ~ shark_sp_metrics$FSp+shark_sp_metrics$Order)
  # m <- lmer(FSp~ED +  (1|Order), data=shark_sp_metrics,  REML=F)

  FSp_ED <- ggplot(shark_sp_metrics,aes(y=FSp,x=ED2))+
              geom_point(aes(col=ext_prob100))+
              scale_colour_gradientn(colours = c("forestgreen", "orange", "red"))+
              theme_classic()+
              ylab("FSp")+
              annotate(geom="text", 180, y = 0.96, label="rho=0.19")
  # 
  # cor.test(shark_sp_metrics$ED3, shark_sp_metrics$FUS,
  #        method = "spearman")
  # 
  # qplot(ED,FSp, facets = ~ Order, data = shark_sp_metrics)+
  #   geom_smooth(method = lm, se=F)
  # 
  # qplot(ED,FSp, facets = ~ iucn, data = shark_sp_metrics)+
  #   geom_smooth(method = lm, se=F)
  # 
  # FUS_ED <- ggplot(shark_sp_metrics,aes(y=FUS,x=ED))+
  #           geom_point(aes(col=iucn_P100))+
  #             scale_colour_gradientn(colours = c("forestgreen", "orange", "red"))+
  #           theme_classic()+ylab("FUS (FUn+FSp)")+
  #           annotate(geom="text", 180, y = 1.96, label="rho=0.15")
  # 
  # qplot(ED,FUS, facets = ~ Order, data = shark_sp_metrics)+
  #   geom_smooth(method = lm, se=F)
  # 
  # qplot(ED,FUS, facets = ~ iucn, data = shark_sp_metrics)+
  #   geom_smooth(method = lm, se=F)
  
  plot_grid(FUn_ED, FSp_ED, nrow=2, labels = LETTERS [1:2])

  ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/FD_ED.jpg",
  plot = last_plot(),width = 15,height = 18, units = c("cm"), dpi = 300)
  
# find common 20
 intersect(top_species$top_ED2$Species,top_species$top_FUn)
 intersect(top_species$top_ED2$Species,top_species$top_FSp$Species)
 intersect(top_species$top_EDGE2$Species,top_species$top_FUSE$Species)
 
# find common 50 fuse_edge
 ed50 <- shark_sp_metrics %>% slice_max(ED2, n = 50) %>% dplyr:: select(Species)
 fsp50 <- shark_sp_metrics %>% slice_max(FSp, n = 50) %>% dplyr:: select(Species)
 fun50 <- shark_sp_metrics %>% slice_max(FUn, n = 50) %>% dplyr:: select(Species)
 intersect(ed50$Species, fsp50 $Species)
 intersect(ed50$Species, fun50 $Species)

 edge50 <- shark_sp_metrics %>% slice_max(EDGE2, n = 50) %>% dplyr:: select(Species)
 fuse50 <- shark_sp_metrics %>% slice_max(FUSE, n = 50) %>% dplyr:: select(Species)
 intersect(edge50$Species, fuse50$Species)
 
# find uncertainty around FUSE based on imputations
  uncert <- readRDS("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_inputs/fuse_uncertainty.rds")
  
  FUSE_sp <- as_tibble(FUSE_sp) %>% 
    left_join(uncert, by = "Species") %>% 
    mutate(diff=FUSE-fuse_mean)
  
  FUSE_sp %>%
    summarise(max_diff= max(diff),
              max_fuse=max(FUSE),
              min_fuse=min(FUSE),
              max_fuse_mean=max(fuse_mean),
              min_fuse_mean=min(fuse_mean),
              max_sd=max(fuse_sd),
              mean_sd=mean(fuse_sd))
  
  FUSE_sp %>% subset(diff>0) %>%
    summarise(sp_dif= n_distinct(Species))

  setdiff(
    FUSE_sp %>% slice_max(FUSE, n = 50) %>% dplyr::select(Species),
    FUSE_sp %>% slice_max(fuse_mean, n = 50) %>% dplyr::select(Species)
  )


  top_FUn <- 
    ggplot(top_species, aes(x=reorder(top_FUn, FUn), y=FUn, color=iucn_status, fill=iucn_status, label=NA)) +
    geom_point(size = 4) + 
    scale_color_manual(values=iucn.cols)+
    geom_segment(aes(x=top_FUn, xend=top_FUn, y=0, yend=FUn))+
    coord_flip()+
    xlab(NULL)+
    ylab("FUn")+
    #geom_text(color = "black", size = 3)+
    theme_bw()+
    theme_classic()+
    theme( axis.text.y = element_text(face="italic"))
  
  top_FSp <- 
    ggplot(top_species, aes(x=reorder(top_FSp$Species, top_FSp$FSp), y=top_FSp$FSp, 
                            color=top_FSp$iucn_status, fill=top_FSp$iucn_status, label=NA)) +
    geom_point(size = 4) + 
    scale_color_manual(values=iucn.cols)+
    geom_segment(aes(x=top_FSp$Species, xend=top_FSp$Species, y=0, yend=top_FSp$FSp))+
    coord_flip()+
    xlab(NULL)+
    ylab("FSp")+
    #geom_text(color = "black", size = 3)+
    theme_bw()+
    theme_classic()+
    theme(legend.position = "none", axis.text.y = element_text(face="italic"))
  
  top_ED2 <- 
    ggplot(top_species, aes(x=reorder(top_ED2$Species, top_ED2$ED2), y=top_ED2$ED2, 
                            color=top_ED2$iucn_status, fill=top_ED2$iucn_status, label=NA)) +
    geom_point(size = 4) + 
    scale_color_manual(values=iucn.cols)+
    geom_segment(aes(x=top_ED2$Species, xend=top_ED2$Species, y=0, yend=top_ED2$ED2))+
    coord_flip()+
    xlab(NULL)+
    ylab("ED2")+
    #geom_text(color = "black", size = 3)+
    theme_bw()+
    theme_classic()+
    theme(legend.position = "none", axis.text.y = element_text(face="italic"))
  
 top_FUSE <- 
    ggplot(top_species, aes(x=reorder(top_FUSE$Species, top_FUSE$FUSE), y=top_FUSE$FUSE, 
                            color=top_FUSE$iucn_status, fill=top_FUSE$iucn_status, label=NA)) +
    geom_point(size = 4) + 
    scale_color_manual(values=iucn.cols)+
    geom_segment(aes(x=top_FUSE$Species, xend=top_FUSE$Species, y=0, yend=top_FUSE$FUSE))+
    coord_flip()+
    xlab(NULL)+
    ylab("FUSE")+
    geom_text(color = "black", size = 3)+
    theme_bw()+
    theme_classic()+
    theme(axis.text.y = element_text(face="italic"), legend.position = "none")
 
 top_FUSE_sd <- 
   ggplot(FUSE_sp %>% 
           slice_max(FUSE, n = 20),
         aes(x=Species,y=fuse_sd))+
    geom_segment(aes(x=reorder(Species, FUSE), xend=reorder(Species, FUSE), y=0, yend=fuse_sd))+
   scale_color_manual(values=iucn.cols)+
    coord_flip()+
    xlab(NULL)+
    ylab("FUSE_sd")+
    #geom_text(color = "black", size = 3)+
    theme_bw()+
    theme_classic()+
    theme( axis.text.y = element_text(face="italic"))

  top_EDGE2 <- 
    ggplot(top_species, aes(x=reorder(top_EDGE2$Species, top_EDGE2$EDGE2), y=top_EDGE2$EDGE2, 
                            color=top_EDGE2$iucn_status, fill=top_EDGE2$iucn_status, label=NA)) +
    geom_point(size = 4) + 
    scale_color_manual(values=iucn.cols)+
    geom_segment(aes(x=top_EDGE2$Species, xend=top_EDGE2$Species, y=0, yend=top_EDGE2$EDGE2))+
    coord_flip()+
    xlab(NULL)+
    ylab("EDGE2")+
    #geom_text(color = "black", size = 3)+
    theme_bw()+
    theme_classic()+
    theme(legend.position = "none", axis.text.y = element_text(face="italic"))
  
    plot_grid(top_FUn, top_FSp, top_ED2, top_FUSE, top_FUSE_sd, top_EDGE2, nrow=2, labels= LETTERS[1:6])

  # save 
    ggsave("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/top20.jpg",
    plot = last_plot(),width = 55,height = 30, units = c("cm"), dpi = 300)
    
    pdf("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/top20.pdf",width=30, height=30)
    plot_grid(top_FUn, top_FSp, top_ED2, top_FUSE, top_FUSE_sd, top_EDGE2, nrow=2)
    dev.off()
    
# finding top 25% FUSE, FUn and FSp (by Fabien Leprieur)
  FUSE_prob100<-shark_sp_metrics
  FUSE_prob100$Species <- gsub(" ", "_", FUSE_prob100$Species)

  Top_FUSEP100_Q25<-FUSE_prob100[FUSE_prob100$FUSE_P100>quantile(FUSE_prob100$FUSE_P100, 0.75),]
  saveRDS(Top_FUSEP100_Q25, file="Top_FUSEP100_Q25_final.rds")

  Top_FUn_Q25<-FUSE_prob100[FUSE_prob100$FUn_std>quantile(FUSE_prob100$FUn_std, 0.75),]
  saveRDS(Top_FUn_Q25, file="Top_FUn_Q25_final.rds")
  
  Top_FSp_Q25<-FUSE_prob100[FUSE_prob100$FSp_std>quantile(FUSE_prob100$FSp_std, 0.75),]
  saveRDS(Top_FSp_Q25, file="Top_FSp_Q25_final.rds")
  
```

# 9. Fig. 1
```{r chunk9, message=FALSE,warning=FALSE}


fig1AB <- plot_grid(vol1.sharks, vol1.Carcharhiniformes, vol1.Lamniformes, vol1.Orectolobiformes, vol1.Heterodontiformes,
              vol1.batoids,vol1.Squaliformes, vol1.Pristiophoriformes,vol1.Squatiniformes, vol1.Hexanchiformes,
              vol1.threat_paper,vol1.Myliobatiformes,vol1.Rhinopristiformes, vol1.Torpediniformes, vol1.Rajiformes,
              nrow=3)

  top12_FUSE <- 
      ggplot(top_species[1:12,], aes(x=reorder(top_FUSE$Species, top_FUSE$FUSE), y=top_FUSE$FUSE, 
                              color=top_FUSE$iucn_status, fill=top_FUSE$iucn_status, label=NA)) +
      geom_point(size = 4) + 
      scale_color_manual(values=iucn.cols)+
      geom_segment(aes(x=top_FUSE$Species[1:12], xend=top_FUSE$Species[1:12], y=0, yend=top_FUSE$FUSE[1:12]))+
      coord_flip()+
      xlab(NULL)+
      ylab("FUSE_P100")+
      geom_text(color = "black", size = 3)+
      theme_bw()+
      theme_classic()+
      theme(axis.text.y = element_text(face="italic"),legend.position = "none",)

fig1CD <-  plot_grid(top12_FUSE, FUn_orders, FSp_orders, FUSE_orders, nrow=1)


  pdf("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/Plots/Figure1_ms.pdf", 
  width = 15,height = 14)

  plot_grid(fig1AB, fig1CD, ncol=1)
 
    dev.off()
```

# 10. Calculate Fric, FSp, Fun and Species richness per grid
```{r chunk10, message=FALSE,warning=FALSE}
 library(reshape2)
 source("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Functions/fonction_FRIC_Global_full.distr.R")
 source("~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/Functions/get_indicator_function 2.R")
 attributes(distrait)$Labels<-colnames(grids.fd) 
 
 # FRic per grid cell  
    test.fd <- apply(grids.fd,1,function(x){colnames(grids.fd)[which(x==1)]})
    grids.summary <- as.data.frame(lengths(test.fd))
    min(grids.summary$`lengths(test.fd)`)

    lala<-lapply(test.fd, function(y){get_FV_Sp (ax=c(1:3), pcoa=pcoa, Selected_sp=y)}); beep()
    
  # extract Fric and S per cell
     s <- unlist(lapply(lala, function(k){
    return(k$data$RS)
  }))
    
     Fric <- unlist(lapply(lala, function(k){
    return(k$data$Fric_r)
  }))
     
    FSp <- unlist(lapply(lala, function(k){
    return(k$specialization[1])
  }))
    
  # create a data.frame with lat, lon and Fric 
    Fric <- as.data.frame(cbind(s,Fric, FSp)) %>%
      rownames_to_column("grid") 
    
    Fric$lat <- sapply(strsplit(Fric$grid, "_"),"[",1)
    Fric$lon <- sapply(strsplit(Fric$grid, "_"),"[",2)
    Fric<- Fric [, c(1, 5, 6, 2, 3, 4)]
  
save(Fric,file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/FRic_grids.RData')
#load('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/FRic_grids.RData')

# FUn per grid cell
  # convert distrait as matrix
    distrait.uni<-as.matrix(distrait)

 # uniquneness using 1 neighbours
   test.uni1 <- apply(grids.fd,1,function(x){colnames(grids.fd)[which(x==1)]}) 
   lala.uni1<-lapply(test.uni1, function(y){get_FU_region (Mat_dist=distrait.uni, nb_NN=1, Selected_sp=y)}); beep()
   
  # store data in data.frame
    uni1<-do.call(rbind, lala.uni1) %>%
      rownames_to_column("grid") 
    
    uni1$lat <- sapply(strsplit(uni1$grid, "_"),"[",1)
    uni1$lon <- sapply(strsplit(uni1$grid, "_"),"[",2)
    uni1<- uni1 [, c(1, 4, 5, 2)]# *note there are some NAs
    
    plot(rasterFromXYZ(data.frame (uni1$lat, uni1$lon, uni1$FU_Region_Mean)), 
       col=rev(heat.colors(10)), main="Uniquness 1NN per grid")
    
save(uni1,file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/uni1.RData')
#load('~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/uni1.RData')

```

# 11. Mean FUn_std, FSp_std, FUSE (global, species-based metrics, mean values per grid)
```{r chunk11, message=FALSE,warning=FALSE}
library(dplyr)
# get IUCN per grid
  grid.list <- apply(grids.fd,1,function(x){colnames(grids.fd)[which(x==1)]}) 
  
  grids.list <- lapply(grid.list, function(k){
               as.data.frame(k)})
  
  means.global <- shark_sp_metrics %>%
    dplyr::select(Species, FUn, FSp, FUSE, ED2, EDGE2, ext_prob100) %>%
    dplyr::rename(k = Species)
   
   grids.mean.sp.metrics <- lapply(grids.list, function(k){
                 dplyr::left_join(k, means.global, by="k")})
  
   save(uni1,file='~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/grids.mean.sp.metrics.RData')

# FUn
   FUn.grids <- lapply(grids.mean.sp.metrics , function(k){
                 return(rbind(mean(k$FUn, na.rm=TRUE)))})
   FUn.grids <- as.data.frame(unlist(FUn.grids))
   FUn.grids$lat <- sapply(strsplit(rownames(FUn.grids), "_"),"[",1)
   FUn.grids$lon <- sapply(strsplit(rownames(FUn.grids), "_"),"[",2)
   FUn.grids<- FUn.grids [, c(2, 3, 1)]
    
    mean_global_FUn_grids<- FUn.grids %>%
    rownames_to_column("grid") %>%
    dplyr::rename(mean_global_FUn_grids = `unlist(FUn.grids)`) 
    
# FSp
  FSp.grids <- lapply(grids.mean.sp.metrics , function(k){
                 return(rbind(mean(k$FSp, na.rm=TRUE)))})
   FSp.grids <- as.data.frame(unlist(FSp.grids))
   FSp.grids$lat <- sapply(strsplit(rownames(FSp.grids), "_"),"[",1)
   FSp.grids$lon <- sapply(strsplit(rownames(FSp.grids), "_"),"[",2)
   FSp.grids<- FSp.grids [, c(2, 3, 1)]
    
    mean_global_FSp_grids<- FSp.grids %>%
    rownames_to_column("grid") %>%
    dplyr::rename(mean_global_FSp_gridss = `unlist(FSp.grids)`) 
    
# GE
  GE.grids <- lapply(grids.mean.sp.metrics , function(k){
                 return(rbind(mean(k$ext_prob100, na.rm=TRUE)))})
   GE.grids <- as.data.frame(unlist(GE.grids))
   GE.grids$lat <- sapply(strsplit(rownames(GE.grids), "_"),"[",1)
   GE.grids$lon <- sapply(strsplit(rownames(GE.grids), "_"),"[",2)
   GE.grids<- GE.grids [, c(2, 3, 1)]
    
   mean_global_GE_grids<- GE.grids %>%
    rownames_to_column("grid") %>%
    dplyr::rename(mean_global_GE_grids = `unlist(GE.grids)`) 

# FUSE
  FUSE.grids <- lapply(grids.mean.sp.metrics , function(k){
                 return(rbind(mean(k$FUSE, na.rm=TRUE)))})
   FUSE.grids <- as.data.frame(unlist(FUSE.grids))
   FUSE.grids$lat <- sapply(strsplit(rownames(FUSE.grids), "_"),"[",1)
   FUSE.grids$lon <- sapply(strsplit(rownames(FUSE.grids), "_"),"[",2)
   FUSE.grids<- FUSE.grids [, c(2, 3, 1)]
    
   mean_global_FUSE_grids<- FUSE.grids %>%
    rownames_to_column("grid") %>%
    dplyr::rename(mean_global_FUSE_grids = `unlist(FUSE.grids)`) 
    
```
  
# 12. Combine all data for spatial quantitative analyses
```{r chunk12, message=FALSE,warning=FALSE}  
library(raster)
library(rgdal)

fd.metrics.grid <- Fric %>%
  left_join(uni1[,c(1,4)], by="grid") %>%
  dplyr::rename(FUn_inside_cells_1NN=FU_Region_Mean) %>%
  dplyr::rename(FSp_inside_cells=FSp) %>%
  # left_join(uni10[,c(1,4)], by="grid") %>%
  # dplyr::rename(uni10=FU_Region_Mean) %>%
  #left_join(grids.uni5) %>%
  # left_join(grids.uni10) %>%
  left_join(mean_global_FUn_grids[,c(1,4)], by="grid") %>%
  left_join(mean_global_FSp_grids[,c(1,4)], by="grid") %>%
  left_join(mean_global_GE_grids[,c(1,4)], by="grid") %>%
  left_join(mean_global_FUSE_grids[,c(1,4)], by="grid")

head(fd.metrics.grid)

write.csv(fd.metrics.grid, file="~/Dropbox (Smithsonian)/SHARKS/SHARKS-TR/Sharks FD/Analyses/R_outputs/fd.metrics.grid.csv")
```
